requirements.txt
Django>=1.6
South==0.8.4
psycopg2==2.5.2
gunicorn==19.1.0
wsgiref==0.1.2
dj-static==0.0.6
dj-database-url==0.3.0
requests==2.2.1
python-twitter==1.3.1
unbabel-py==0.34

